requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyio==4.8.0
argcomplete==3.6.3
attrs==25.4.0
bcc==0.35.0
bcrypt==4.3.0
Beaker==1.12.1
beautifulsoup4==4.14.2
blivet==3.12.1
blivet-gui==2.6.0
boto3==1.40.70
botocore==1.40.70
Brlapi==0.8.7
Brotli==1.1.0
certifi==2025.7.9
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.7
cockpit @ file:///builddir/build/BUILD/cockpit-351-build/cockpit-351/tmp/wheel/cockpit-351-py3-none-any.whl#sha256=59d2dcd42c86f8d88df2a0b4afd6bee5c58753efde42f78d6ed1686b14aa61df
crypt_r==3.13.1
cryptography==45.0.4
cupshelpers==1.0
dasbus==1.7
dbus-fast==2.44.3
dbus-python==1.4.0
dbus_next==0.2.3
distro==1.9.0
dnf==4.24.0
et_xmlfile==2.0.0
fedora-third-party==0.10
fido2==2.0.0
file-magic==0.4.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.13.0
i3ipc==2.2.1
idna==3.10
importlib_metadata==8.7.0
iso639==0.1.4
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.6.0
langtable==0.0.69
libcomps==0.1.23
libdnf==0.75.0
libdnf5==5.2.17.0
libvirt-python==11.6.0
louis==3.33.0
lxml==6.0.1
Mako==1.2.3
MarkupSafe==3.0.2
mercurial==7.1.1
more-itertools==10.5.0
multidict==6.6.4
netaddr==1.3.0
netifaces==0.11.0
nftables==0.1
nwg-panel==0.10.12
olefile==0.47
openpyxl==3.1.5
packaging==25.0
Paste==3.10.1
perf==0.1
pexpect==4.9.0
pid==3.0.4
pillow==11.3.0
ply==3.11
productmd==1.48
progressbar2==4.5.0
propcache==0.4.1
proton-core==0.7.3
proton-keyring-linux==0.2.1
proton-vpn-api-core==4.13.1
proton-vpn-daemon==0.13.4
proton-vpn-gtk-app==4.12.0
proton-vpn-network-manager==0.13.4
psutil==7.0.0
ptyprocess==0.7.0
pwquality==1.4.5
PyAudio==0.2.13
pycairo==1.28.0
pycares==4.10.0
pycparser==2.22
pycrypto==2.6.1
pycups==2.0.4
pyenchant==3.2.2
PyGObject==3.54.5
pyinotify==0.9.6
pykickstart @ file:///builddir/build/BUILD/pykickstart-3.66-build/pykickstart-3.66
PyNaCl==1.5.0
pyOpenSSL==25.1.0
pyparted==3.13.0
PySocks==1.7.1
python-augeas==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-gnupg==0.5.4
python-linux-procfs==0.7.3
python-meh==0.52
python-pam==2.0.2
python-utils==3.9.1
python-xlib==0.33
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.2
pyynl @ file:///builddir/build/BUILD/kernel-6.17.7-build/kernel-6.17.7/linux-6.17.7-300.fc43.x86_64/tools/net/ynl
RapidFuzz==3.12.2
referencing==0.36.2
regex==2025.10.23
requests==2.32.5
requests-file==2.0.0
requests-ftp==0.3.1
rpds-py==0.27.0
rpm==6.0.0
rpmautospec==0.8.3
rpmautospec-core==0.1.5
s3transfer==0.14.0
SecretStorage==3.3.3
selinux @ file:///builddir/build/BUILD/libselinux-3.9-build/libselinux-3.9/src
sentry-sdk==2.35.0
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.9-build/selinux-3.9/python/sepolicy
setools==4.6.0
setuptools==78.1.1
shtab==1.7.2
simpleaudio==1.0.4
simpleline==1.9.0
six==1.17.0
sniffio==1.3.1
sos==4.10.1
soupsieve==2.8
systemd-python==235
typing_extensions==4.15.0
ufw==0.35
urllib3==2.5.0
xkbregistry==0.3
yarl==1.22.0
zipp==3.23.0
